Re-arrange multiscene tests into three modules in a sub-directory
